Help with Jump
Hi Joe,

I'm trying to get the jump to be more punctual when sprinting. At the moment, the run jump and walk jumps are the same speed and arc.

I also have an issue with shooting the projectile - it seems to shoot from the right side of the character gameobject and results in the projectile pushing me away.